requirements.txt
matplotlib==3.5.1
numpy==1.23.1
Pillow==9.3.0
scikit_learn==1.2.0
torch==1.12.0+cu116
torchvision==0.13.0+cu116
